move SparkNewHadoopWriter to internal/io
