+++ core/src/main/scala/org/apache/spark/internal/io/SparkNewHadoopWriter.scala
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.spark
+package org.apache.spark.internal.io
 
 import java.util.Date
 
@@ -23,7 +23,7 @@ import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.mapreduce._
 import org.apache.hadoop.mapreduce.task.TaskAttemptContextImpl
 
-import org.apache.spark.internal.io.HadoopMapReduceCommitProtocol
+import org.apache.spark.{SparkHadoopWriter, TaskContext}
 import org.apache.spark.internal.Logging
 import org.apache.spark.util.SerializableConfiguration
